// reference-mcp-servers/Cloud_Platforms/redis_mcp-redis-cloud_015aaa6_typescript/src/clients/generated/models/ActiveActiveSubscriptionRegions.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * List of active-active subscription regions
 */
export type ActiveActiveSubscriptionRegions = {
    subscriptionId?: number;
    links?: Array<Record<string, Record<string, any>>>;
};